lsif -> scip (#42)
rewrote a lot of the tree visitor code. It is now much more
focused on using visitName and getScipSymbol rather than
having some items made in other visitors.

In general, if a visitor does usesful work, then it should return
false and handle all the cases itself (which is now the way that each
of the visitors work). Some future work is still required to untangle
a bit of the complexity for determining how a particular symbol is
created as well as adding some debug capabilities to search for
how a particular symbol is generated.

Also:
- fixed a bunch of bugs with imports and references across files / projects
- added a dockerfile for autoindexing (will be published on next release)
- use `scip` instead of `lsif` for names now
- update `scip` library
- add `external_symbols` support for snapshot formatter (and correspondingly
  the support for actually generated external_symbols for builtin types).
    - In general, we do not generate external_symbols for non-builtin types at this
    time because we trust that those will get autoindexed in the future.