fix: Avoid crash in the presence of decorators. (#184)
Changes the method lookup logic to use the same
logic as Pyright internals instead of our own ad-hoc
implementation of name resolution, which does not
match Python's method resolution order (MRO), which
is based purely on names, and not on types.

This gets rid of a code path where we were doing a `!`
operation, which triggered a crash when indexing
a prospect's codebase.

Fixes GRAPH-1278.